Enable CVO to evaluate conditional updates on self-managed HyperShift deployed on OpenShift
This commit will enable the Cluster Version Operator (CVO) to evaluate
conditional updates on self-managed HyperShift deployed on an OpenShift
management cluster.

For the CVO to evaluate conditional updates, it accesses a PromQL query
service holding the appropriate information. In the standalone
OpenShift, it is the thanos-querier service in the openshift-monitoring
namespace.

This commit will point the hosted CVO at the mentioned service in
self-managed HyperShift running on the OCP management cluster as well.
To use the thanos-querier service, commit utilizes its tenancy port
9092. This port allows the users to query metrics from a specified
namespace (in this case the hosted CVO can query metrics from its hosted
control plane namespace). For the authorization, the hosted CVO is given
a new Role with the minimal needed permissions to be able to query its
HCP metrics.

For the Control Plane Operator to be able to give this permission to
the CVO, it needs this permission as well. The same case applies to
the HyperShift Operator (HSO). Permission is added to the HSO in all
self-managed use cases (where RHOBS monitoring is disabled) because the
HSO is isolated from other components and thus shows a minimal risk of
exploiting this permission. An alternative would be to discover
OpenShift resources before creating and applying HSO manifests, but this
would introduce a need for a Kubernetes discovery client in the
`hypershift install render...` command. Another alternative would be to
create a new flag to explicitly specify whether HyperShift is deployed
on an OpenShift management cluster. The chosen solution doesn't create
new permissions for managed HyperShift while not complicating running
HyperShift in self-managed.

To dynamically discover whether a HyperShift is deployed on an OpenShift
management cluster, a check for route capability is made. This is
inspired by the same logic used in the Grafana Operator for detecting
OpenShift clusters [1].

A new network policy is created to only allow egress communication to
the PromQL query service pod's port from the hosted CVO pods.

[1] https://github.com/grafana-operator/grafana-operator/blob/6d19d89492e910d38aff41aa4bdfa5541a0f3a1f/controllers/autodetect/main.go#L36-L50